requirements.txt


streamlit
altair==5.4.1
certifi==2024.8.30
charset-normalizer==3.4.0
db-dtypes==1.3.0
GitPython==3.1.43
google-api-core==2.21.0
google-auth==2.35.0
google-cloud==0.34.0
google-cloud-bigquery==3.26.0
google-cloud-bigquery-storage==2.26.0
google-cloud-core==2.4.1

google-pasta==0.2.0
seaborn
pandas
numpy